requirements.txt
accelerate
av
bert_score
decord
deepspeed
datasets
flash-attn
gdown
huggingface_hub
matplotlib
nltk
open_clip_torch
openai
opencv-python
peft
pycocoevalcap
rouge_score
scikit-learn
scipy
sentence_transformers
sentencepiece
tabulate
timm
transformers
yt_dlp
wandb